Streamline code, fix bugs, add tests, improve what AI did

+++ backend/WeightTracker.Api/DTOs/UserSettingsDto.cs
@@ -2,9 +2,9 @@ using System.ComponentModel.DataAnnotations;
 
 namespace WeightTracker.Api.DTOs;
 
-public record UserSettingsDto(decimal? HeightCm, string PreferredUnit, decimal? TdeeKcal);
+public record UserSettingsDto(int? HeightCm, string PreferredUnit, int? TdeeKcal);
 
 public record UpdateUserSettingsDto(
-    [Range(50, 300)] decimal? HeightCm,
+    [Range(50, 300)] int? HeightCm,
     [RegularExpression("^(kg|lbs)$")] string PreferredUnit,
-    [Range(1, 15000)] decimal? TdeeKcal);
+    [Range(1, 15000)] int? TdeeKcal);
